Add SFLLL spec and update utils
