--- frontend/tests/e2e/apply/fixtures/sfLLL-fill-data.ts
@@ -1,341 +1,0 @@
-import { FormFillFieldDefinitions } from "tests/e2e/utils/common/types";
-import { FORM_DEFAULTS } from "tests/e2e/utils/forms/form-defaults";
-
-// Regex matcher for SF-LLL: matches both "SF-LLL" and "Disclosure of Lobbying Activities"
-export const SFLLL_FORM_MATCHER =
-  /SF\s*[-‑–—]?\s*LLL|Disclosure\s+of\s+Lobbying\s+Activities/i;
-
-const fieldDefinitionsSFLLL: FormFillFieldDefinitions = {
-  federalAction_type: {
-    selector: "#federal_action_type",
-    type: "dropdown",
-    section: "Section 1",
-    field: "Type of Federal Action",
-  },
-  federalAction_status: {
-    selector: "#federal_action_status",
-    type: "dropdown",
-    section: "Section 2",
-    field: "Status of Federal Action",
-  },
-  federalAction_reportType: {
-    selector: "#report_type",
-    type: "dropdown",
-    section: "Section 3",
-    field: "Report Type",
-  },
-  materialChange_year: {
-    testId: "material_change_year",
-    type: "text",
-    section: "Section 3",
-    field: "Material Change Year",
-  },
-  materialChange_quarter: {
-    testId: "material_change_quarter",
-    type: "text",
-    section: "Section 3",
-    field: "Material Change Quarter",
-  },
-  materialChange_lastReportDate: {
-    testId: "last_report_date",
-    type: "text",
-    section: "Section 3",
-    field: "Last Report Date",
-  },
-  // these two are not being filled
-  reportingEntity_type: {
-    selector: "#reporting_entity--entity_type",
-    type: "dropdown",
-    section: "Section 4",
-    field: "Entity Type",
-  },
-  reportingEntity_tier: {
-    testId: "reporting_entity--tier",
-    type: "text",
-    section: "Section 4",
-    field: "Tier",
-  },
-  reportingEntity_orgName: {
-    testId: "reporting_entity--applicant_reporting_entity--organization_name",
-    type: "text",
-    section: "Section 4",
-    field: "Organization Name",
-  },
-  reportingEntity_street1: {
-    testId: "reporting_entity--applicant_reporting_entity--address--street1",
-    type: "text",
-    section: "Section 4",
-    field: "Street 1",
-  },
-  reportingEntity_street2: {
-    testId: "reporting_entity--applicant_reporting_entity--address--street2",
-    type: "text",
-    section: "Section 4",
-    field: "Street 2",
-  },
-  reportingEntity_city: {
-    testId: "reporting_entity--applicant_reporting_entity--address--city",
-    type: "text",
-    section: "Section 4",
-    field: "City",
-  },
-  reportingEntity_state: {
-    selector: "#reporting_entity--applicant_reporting_entity--address--state",
-    type: "dropdown",
-    section: "Section 4",
-    field: "State",
-  },
-  reportingEntity_zip: {
-    testId: "reporting_entity--applicant_reporting_entity--address--zip_code",
-    type: "text",
-    section: "Section 4",
-    field: "Zip Code",
-  },
-  reportingEntity_district: {
-    testId:
-      "reporting_entity--applicant_reporting_entity--congressional_district",
-    type: "text",
-    section: "Section 4",
-    field: "Congressional District",
-  },
-  primeEntity_orgName: {
-    testId: "reporting_entity--prime_reporting_entity--organization_name",
-    type: "text",
-    section: "Section 5",
-    field: "Prime Organization Name",
-  },
-  primeEntity_street1: {
-    testId: "reporting_entity--prime_reporting_entity--address--street1",
-    type: "text",
-    section: "Section 5",
-    field: "Prime Street 1",
-  },
-  primeEntity_street2: {
-    testId: "reporting_entity--prime_reporting_entity--address--street2",
-    type: "text",
-    section: "Section 5",
-    field: "Prime Street 2",
-  },
-  primeEntity_city: {
-    testId: "reporting_entity--prime_reporting_entity--address--city",
-    type: "text",
-    section: "Section 5",
-    field: "Prime City",
-  },
-  primeEntity_state: {
-    selector: "#reporting_entity--prime_reporting_entity--address--state",
-    type: "dropdown",
-    section: "Section 5",
-    field: "Prime State",
-  },
-  primeEntity_zip: {
-    testId: "reporting_entity--prime_reporting_entity--address--zip_code",
-    type: "text",
-    section: "Section 5",
-    field: "Prime Zip Code",
-  },
-  primeEntity_district: {
-    testId: "reporting_entity--prime_reporting_entity--congressional_district",
-    type: "text",
-    section: "Section 5",
-    field: "Prime Congressional District",
-  },
-  federalInfo_agencyDepartment: {
-    testId: "federal_agency_department",
-    type: "text",
-    section: "Section 6",
-    field: "Federal Agency/Department",
-  },
-  federalInfo_name: {
-    testId: "federal_program_name",
-    type: "text",
-    section: "Section 7",
-    field: "Federal Program Name",
-  },
-  federalInfo_assistanceListingNumber: {
-    testId: "assistance_listing_number",
-    type: "text",
-    section: "Section 7",
-    field: "Assistance Listing Number",
-  },
-  federalInfo_actionNumber: {
-    testId: "federal_action_number",
-    type: "text",
-    section: "Section 8",
-    field: "Federal Action Number",
-  },
-  federalInfo_awardAmount: {
-    testId: "award_amount",
-    type: "text",
-    section: "Section 9",
-    field: "Award Amount",
-  },
-  lobbyingRegistrant_firstName: {
-    testId: "lobbying_registrant--individual--first_name",
-    type: "text",
-    section: "Section 10a",
-    field: "First Name",
-  },
-  lobbyingRegistrant_middleName: {
-    testId: "lobbying_registrant--individual--middle_name",
-    type: "text",
-    section: "Section 10a",
-    field: "Middle Name",
-  },
-  lobbyingRegistrant_lastName: {
-    testId: "lobbying_registrant--individual--last_name",
-    type: "text",
-    section: "Section 10a",
-    field: "Last Name",
-  },
-  lobbyingRegistrant_prefix: {
-    testId: "lobbying_registrant--individual--prefix",
-    type: "text",
-    section: "Section 10a",
-    field: "Prefix",
-  },
-  lobbyingRegistrant_suffix: {
-    testId: "lobbying_registrant--individual--suffix",
-    type: "text",
-    section: "Section 10a",
-    field: "Suffix",
-  },
-  lobbyingRegistrant_street1: {
-    testId: "lobbying_registrant--address--street1",
-    type: "text",
-    section: "Section 10a",
-    field: "Street 1",
-  },
-  lobbyingRegistrant_street2: {
-    testId: "lobbying_registrant--address--street2",
-    type: "text",
-    section: "Section 10a",
-    field: "Street 2",
-  },
-  lobbyingRegistrant_city: {
-    testId: "lobbying_registrant--address--city",
-    type: "text",
-    section: "Section 10a",
-    field: "City",
-  },
-  lobbyingRegistrant_state: {
-    selector: "#lobbying_registrant--address--state",
-    type: "dropdown",
-    section: "Section 10a",
-    field: "State",
-  },
-  lobbyingRegistrant_zip: {
-    testId: "lobbying_registrant--address--zip_code",
-    type: "text",
-    section: "Section 10a",
-    field: "Zip Code",
-  },
-  performingService_firstName: {
-    testId: "individual_performing_service--individual--first_name",
-    type: "text",
-    section: "Section 10b",
-    field: "First Name",
-  },
-  performingService_middleName: {
-    testId: "individual_performing_service--individual--middle_name",
-    type: "text",
-    section: "Section 10b",
-    field: "Middle Name",
-  },
-  performingService_lastName: {
-    testId: "individual_performing_service--individual--last_name",
-    type: "text",
-    section: "Section 10b",
-    field: "Last Name",
-  },
-  performingService_prefix: {
-    testId: "individual_performing_service--individual--prefix",
-    type: "text",
-    section: "Section 10b",
-    field: "Prefix",
-  },
-  performingService_suffix: {
-    testId: "individual_performing_service--individual--suffix",
-    type: "text",
-    section: "Section 10b",
-    field: "Suffix",
-  },
-  performingService_street1: {
-    testId: "individual_performing_service--address--street1",
-    type: "text",
-    section: "Section 10b",
-    field: "Street 1",
-  },
-  performingService_street2: {
-    testId: "individual_performing_service--address--street2",
-    type: "text",
-    section: "Section 10b",
-    field: "Street 2",
-  },
-  performingService_city: {
-    testId: "individual_performing_service--address--city",
-    type: "text",
-    section: "Section 10b",
-    field: "City",
-  },
-  performingService_state: {
-    selector: "#individual_performing_service--address--state",
-    type: "dropdown",
-    section: "Section 10b",
-    field: "State",
-  },
-  performingService_zip: {
-    testId: "individual_performing_service--address--zip_code",
-    type: "text",
-    section: "Section 10b",
-    field: "Zip Code",
-  },
-  signature_prefix: {
-    testId: "signature_block--name--prefix",
-    type: "text",
-    section: "Section 11",
-    field: "Prefix",
-  },
-  signature_firstName: {
-    testId: "signature_block--name--first_name",
-    type: "text",
-    section: "Section 11",
-    field: "First Name",
-  },
-  signature_middleName: {
-    testId: "signature_block--name--middle_name",
-    type: "text",
-    section: "Section 11",
-    field: "Middle Name",
-  },
-  signature_lastName: {
-    testId: "signature_block--name--last_name",
-    type: "text",
-    section: "Section 11",
-    field: "Last Name",
-  },
-  signature_suffix: {
-    testId: "signature_block--name--suffix",
-    type: "text",
-    section: "Section 11",
-    field: "Suffix",
-  },
-  signature_title: {
-    testId: "signature_block--title",
-    type: "text",
-    section: "Section 11",
-    field: "Title",
-  },
-  signature_phone: {
-    testId: "signature_block--telephone",
-    type: "text",
-    section: "Section 11",
-    field: "Telephone",
-  },
-};
-
-export const SFLLL_FORM_CONFIG = {
-  ...FORM_DEFAULTS,
-  formName: SFLLL_FORM_MATCHER,
-  fields: fieldDefinitionsSFLLL,
-} as const;
